clean store structure a bit.
