--- client/src/stores/CommonStore.tsx
@@ -1,11 +1,0 @@
-import { observable } from 'mobx';
-
-export enum UserState {
-    login,
-    signup,
-    connected
-}
-
-export class CommonStore {
-    @observable userState: UserState = UserState.login;
-}
