+++ client/src/stores/index.tsx
@@ -1,4 +1,4 @@
-export * from './CommonStore';
+export * from './NowStore';
 export * from './UIStore';
 export * from './SettingStore';
 export * from './SongStore';
